skin_project/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
openai
pinecone-client
passlib[bcrypt]
email-validator
selenium
beautifulsoup4
pandas
playwright
webdriver-manager
pinecone-client
